automatically backed up by learn

+++ index.js
@@ -20,5 +20,5 @@ function makeInt(n) {
   return parseInt(n, 10);
 }
 function preserveDecimal(n) {
-  return parseInt('2.222', 10);
+  return parseFloat('2.222', 10);
 }
